#155 - reorganize components, fix cell voltage decimals, trick out faults component
