--- angular-client/src/components/battery-components/fault-display/fault-display.component.html
@@ -1,11 +1,0 @@
-<info-background title="Faults" icon="warning_amber">
-  <div style="min-width: 20vw; padding-right: 30px" class="fault-text-body">
-    <vstack>
-      <div style="height: 25px"></div>
-      <ng-container *ngFor="let fault of getFaultStrings()">
-        <typography variant="info-subtitle" [content]="fault"></typography>
-      </ng-container>
-      <div style="height: 50px"></div>
-    </vstack>
-  </div>
-</info-background>
