+++ angular-client/src/pages/charging-page/charging-page.component.html
@@ -19,26 +19,26 @@
       </div>
 
       <hstack style="width: 100%">
-        <state-of-charge-display style="width: 20vw" />
-        <charging-state style="width: 40%" />
-        <hstack style="width: 40%">
-          <battery-status-display style="width: 100%" />
-          <BMS-mode-display style="width: 100%" />
-        </hstack>
-      </hstack>
-
-      <hstack style="width: 100%">
-        <fault-display style="width: 19.3vw; height: 405px" />
-        <vstack style="width: 81vw; height: 400px">
-          <cell-temp-display style="width: 100%; height: 200px" />
-          <high-low-cell-display style="width: 100%; height: 200px" />
+        <vstack style="width: 20%" spacing="15px">
+          <state-of-charge-display style="height: 142px; width: 100%" />
+          <fault-display style="height: 580px; width: 100%" />
         </vstack>
-      </hstack>
-
-      <hstack style="width: 100%">
-        <current-display style="height: 150px; width: 15%" />
-        <pack-temp style="height: 150px; width: 15%" />
-        <pack-voltage-display style="height: 150px; width: 70%" />
+        <vstack style="width: 80%" spacing="15px">
+          <hstack style="width: 100%">
+            <charging-state style="width: 25%" />
+            <battery-status-display style="width: 25%" />
+            <BMS-mode-display style="width: 25%" />
+            <pack-temp style="height: 142px; width: 25%" />
+          </hstack>
+          <div style="width: 100%; height: 580px; display: flex; flex-direction: column; justify-content: space-between">
+            <cell-temp-display style="width: 100%; height: 200px" />
+            <high-low-cell-display style="width: 100%; height: 200px" />
+            <hstack style="width: 100%">
+              <pack-voltage-display style="height: 150px; width: 80%" />
+              <current-display style="height: 150px; width: 20%" />
+            </hstack>
+          </div>
+        </vstack>
       </hstack>
     </vstack>
   </ng-template>
