+++ angular-client/src/pages/charging-page/components/BMS-mode/BMS-mode-display.component.ts
