+++ angular-client/src/pages/charging-page/components/battery-current/current-display/current-display.component.html
