+++ angular-client/src/pages/charging-page/components/battery-info-display/battery-info-desktop/battery-info-desktop.component.css
