+++ angular-client/src/pages/charging-page/components/battery-info-display/battery-info-display.html
