+++ angular-client/src/pages/charging-page/components/battery-info-display/battery-info-mobile/battery-info-mobile.component.css
