+++ angular-client/src/pages/charging-page/components/battery-percentage/battery-percentage.component.html
