+++ angular-client/src/pages/charging-page/components/battery-status-display/battery-status-display.component.html
