+++ angular-client/src/pages/charging-page/components/cell-temp/cell-temp-display/cell-temp-display.component.ts
