+++ angular-client/src/pages/charging-page/components/cell-temp/cell-temp-graph/cell-temp-graph.component.html
