+++ angular-client/src/pages/charging-page/components/charging-state/charging-state.component.html
