+++ angular-client/src/pages/charging-page/components/fault-display/fault-display.component.css
@@ -1,7 +1,8 @@
 .fault-text-body {
-  height: 80%;
+  height: 95%;
   display: flex;
   align-items: left;
   justify-content: left;
   padding-left: 10%;
+  overflow: hidden;
 }
