+++ angular-client/src/pages/charging-page/components/fault-display/fault-display.component.html
@@ -1,0 +1,21 @@
+<info-background title="Faults" icon="warning_amber">
+  <div style="min-width: 20vw; padding-right: 30px; padding-left: 0px" class="fault-text-body">
+    <vstack style="width: 100%">
+      <div style="height: 10px"></div>
+      <ng-container *ngFor="let fault of this.faults.reverse(); index as i">
+        <div
+          [style]="
+            i % 2 == 0
+              ? 'background-color: #d4760b	; width: 100%; border-radius: 2px'
+              : 'background-color: #4d453c; width: 100%; border-radius: 2px'
+          "
+        >
+          <typography variant="info-subtitle" [content]="fault.faultName"></typography>
+          <typography variant="info-subtitle" [content]="fault.time"></typography>
+        </div>
+        <div style="height: 2px"></div>
+      </ng-container>
+      <div style="height: 50px"></div>
+    </vstack>
+  </div>
+</info-background>
