+++ angular-client/src/pages/charging-page/components/high-low-cell/high-low-cell-display/high-low-cell-display.component.css
