+++ angular-client/src/pages/charging-page/components/high-low-cell/high-low-cell-display/high-low-cell-display.component.html
@@ -11,7 +11,7 @@
     <divider style="height: 100px"></divider>
 
     <vstack style="width: 10%">
-      <typography variant="info-value" [content]="lowCellVolts.toString() + 'V'" additionalStyles="fontSize:30px" />
+      <typography variant="info-value" [content]="lowCellVoltage.toString() + 'V'" additionalStyles="fontSize:30px" />
       <typography variant="info-subtitle" content="Low Cell" additionalStyles="fontSize:15px" />
     </vstack>
 
@@ -20,7 +20,7 @@
 
     <vstack style="width: 10%">
       <hstack>
-        <typography variant="info-value" [content]="highCellVolts.toString() + 'V'" additionalStyles="fontSize:30px" />
+        <typography variant="info-value" [content]="highCellVoltage.toString() + 'V'" additionalStyles="fontSize:30px" />
       </hstack>
       <typography variant="info-subtitle" content="High Cell" additionalStyles="fontSize:15px" />
     </vstack>
