+++ angular-client/src/pages/charging-page/components/high-low-cell/high-low-cell-display/high-low-cell-display.component.ts
@@ -10,8 +10,8 @@ import { decimalPipe } from 'src/utils/pipes.utils';
 })
 export default class HighLowCellDisplay {
   delta: number = 0;
-  lowCellVolts: number = 0;
-  highCellVolts: number = 0;
+  lowCellVoltage: number = 0;
+  highCellVoltage: number = 0;
   mobileThreshold = 1000;
   isDesktop = window.innerWidth > this.mobileThreshold;
 
@@ -23,13 +23,13 @@ export default class HighLowCellDisplay {
   }
 
   ngOnInit() {
-    this.storage.get(IdentifierDataType.VOLTS_HIGH).subscribe((value) => {
-      this.lowCellVolts = decimalPipe(value.values[0]);
-      this.delta = decimalPipe((this.highCellVolts - this.lowCellVolts).toFixed(1));
-    });
     this.storage.get(IdentifierDataType.VOLTS_LOW).subscribe((value) => {
-      this.highCellVolts = decimalPipe(value.values[0]);
-      this.delta = decimalPipe((this.highCellVolts - this.lowCellVolts).toFixed(2));
+      this.lowCellVoltage = decimalPipe(value.values[0], 3);
+      this.delta = decimalPipe((this.highCellVoltage - this.lowCellVoltage).toFixed(3), 3);
+    });
+    this.storage.get(IdentifierDataType.VOLTS_HIGH).subscribe((value) => {
+      this.highCellVoltage = decimalPipe(value.values[0], 3);
+      this.delta = decimalPipe((this.highCellVoltage - this.lowCellVoltage).toFixed(3), 3);
     });
   }
 }
