+++ angular-client/src/pages/charging-page/components/pack-temp/pack-temp.component.html
