+++ angular-client/src/pages/charging-page/components/state-of-charge/state-of-charge-display/state-of-charge-display.component.css
